EZP-24704: introduce AbstractSubtree search engine slot
